+++ src/Microsoft.Health.Dicom.Operations.Functions.UnitTests/Microsoft.Health.Dicom.Operations.Functions.UnitTests.csproj
@@ -13,8 +13,14 @@
 
   <ItemGroup>
     <ProjectReference Include="..\Microsoft.Health.Dicom.Core\Microsoft.Health.Dicom.Core.csproj" />
-    <ProjectReference Include="..\Microsoft.Health.Dicom.Functions\Microsoft.Health.Dicom.Functions.csproj" />
+    <ProjectReference Include="..\Microsoft.Health.Dicom.Operations.Functions\Microsoft.Health.Dicom.Operations.Functions.csproj" />
     <ProjectReference Include="..\Microsoft.Health.Dicom.Tests.Common\Microsoft.Health.Dicom.Tests.Common.csproj" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <None Include="Indexing\ReindexDurableFunctionTests.*.cs">
+      <DependentUpon>ReindexDurableFunctionTests.cs</DependentUpon>
+    </None>
   </ItemGroup>
 
 </Project>
