+++ src/Microsoft.Health.Dicom.Operations.Functions/Indexing/Models/PrepareReindexingTagsInput.cs
@@ -3,21 +3,24 @@
 // Licensed under the MIT License (MIT). See LICENSE in the repo root for license information.
 // -------------------------------------------------------------------------------------------------
 
-namespace Microsoft.Health.Dicom.Functions.Indexing.Models
+using System.Collections.Generic;
+
+namespace Microsoft.Health.Dicom.Operations.Functions.Indexing.Models
 {
     /// <summary>
-    /// Represents input to activity <see cref="ReindexDurableFunction.UpdateReindexProgressAsync"/>.
+    ///  Represents input to <see cref="ReindexDurableFunction.PrepareReindexingTagsAsync"/>
     /// </summary>
-    public class UpdateReindexProgressInput
+    public class PrepareReindexingTagsInput
     {
         /// <summary>
-        /// Gets or sets the end watermark of operation.
+        /// Gets or sets the operation id
         /// </summary>
-        public long EndWatermark { get; set; }
-
-        ///// <summary>
-        ///// Gets or sets the operation id.
-        ///// </summary>
         public string OperationId { get; set; }
+
+        /// <summary>
+        /// Gets or sets the tag keys
+        /// </summary>
+        public IReadOnlyList<int> TagKeys { get; set; }
+
     }
 }
