+++ src/Microsoft.Health.Dicom.Operations.Functions/Indexing/ReindexDurableFunction.Orchestration.cs
@@ -13,9 +13,9 @@ using Microsoft.Azure.WebJobs.Extensions.DurableTask;
 using Microsoft.Extensions.Logging;
 using Microsoft.Health.Dicom.Core.Features.ExtendedQueryTag;
 using Microsoft.Health.Dicom.Core.Features.Indexing;
-using Microsoft.Health.Dicom.Functions.Indexing.Models;
+using Microsoft.Health.Dicom.Operations.Functions.Indexing.Models;
 
-namespace Microsoft.Health.Dicom.Functions.Indexing
+namespace Microsoft.Health.Dicom.Operations.Functions.Indexing
 {
     public partial class ReindexDurableFunction
     {
