+++ src/Microsoft.Health.Dicom.Operations.Functions/Indexing/ReindexDurableFunction.cs
@@ -1,0 +1,50 @@
+// -------------------------------------------------------------------------------------------------
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License (MIT). See LICENSE in the repo root for license information.
+// -------------------------------------------------------------------------------------------------
+
+using EnsureThat;
+using Microsoft.Extensions.Options;
+using Microsoft.Health.Dicom.Core.Features.ExtendedQueryTag;
+using Microsoft.Health.Dicom.Core.Features.Indexing;
+using Microsoft.Health.Dicom.Core.Features.Retrieve;
+using Microsoft.Health.Dicom.Operations.Functions.Indexing.Configuration;
+
+namespace Microsoft.Health.Dicom.Operations.Functions.Indexing
+{
+    /// <summary>
+    /// Represents the Azure Durable Functions that perform the re-indexing of previously added DICOM instances
+    /// based on new tags configured by the user.
+    /// </summary>
+    public partial class ReindexDurableFunction
+    {
+        private readonly ReindexConfiguration _reindexConfig;
+        private readonly IReindexStore _reindexStore;
+        private readonly IInstanceReindexer _instanceReindexer;
+        private readonly IAddExtendedQueryTagService _addExtendedQueryTagService;
+        private readonly IInstanceStore _instanceStore;
+        private readonly IExtendedQueryTagStore _extendedQueryTagStore;
+
+        public ReindexDurableFunction(
+            IOptions<IndexingConfiguration> configOptions,
+            IAddExtendedQueryTagService addExtendedQueryTagService,
+            IReindexStore reindexStore,
+            IInstanceStore instanceStore,
+            IInstanceReindexer instanceReindexer,
+            IExtendedQueryTagStore extendedQueryTagStore)
+        {
+            EnsureArg.IsNotNull(configOptions, nameof(configOptions));
+            EnsureArg.IsNotNull(reindexStore, nameof(reindexStore));
+            EnsureArg.IsNotNull(instanceReindexer, nameof(instanceReindexer));
+            EnsureArg.IsNotNull(addExtendedQueryTagService, nameof(addExtendedQueryTagService));
+            EnsureArg.IsNotNull(instanceStore, nameof(instanceStore));
+            EnsureArg.IsNotNull(extendedQueryTagStore, nameof(extendedQueryTagStore));
+            _reindexConfig = configOptions.Value.Add;
+            _reindexStore = reindexStore;
+            _instanceReindexer = instanceReindexer;
+            _addExtendedQueryTagService = addExtendedQueryTagService;
+            _instanceStore = instanceStore;
+            _extendedQueryTagStore = extendedQueryTagStore;
+        }
+    }
+}
