+++ src/Microsoft.Health.Dicom.Operations.Functions/Microsoft.Health.Dicom.Operations.Functions.csproj
@@ -24,6 +24,12 @@
   </ItemGroup>
 
   <ItemGroup>
+    <None Include="Indexing\ReindexDurableFunction.*.cs">
+      <DependentUpon>ReindexDurableFunction.cs</DependentUpon>
+    </None>
+  </ItemGroup>
+
+  <ItemGroup>
     <None Update="host.json">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </None>
