+++ src/Microsoft.Health.Dicom.Operations.Functions/Properties/AssemblyInfo.cs
