+++ src/Microsoft.Health.Dicom.Operations.Functions/Properties/serviceDependencies.json
