+++ src/Microsoft.Health.Dicom.Operations.Functions/Properties/serviceDependencies.local.json
