+++ src/Microsoft.Health.Dicom.Operations.Functions/host.json
