+++ src/Microsoft.Health.Dicom.Operations.Functions/local.settings.json
